requirements.txt
Click==7.0
Flask==1.0.2
gitdb2==2.0.5
GitPython==2.1.11
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
smmap2==2.0.5
Werkzeug==0.15.4
